+++ src/main/groovy/org/codenarc/CodeNarc.groovy
@@ -89,12 +89,12 @@ Usage: java org.codenarc.CodeNarc [OPTIONS]
         The title for this analysis; used in the output report(s), if supported by the report type. Optional.
     -report=<REPORT-TYPE[:FILENAME]>
         The definition of the report to produce. The option value is of the form
-        TYPE[:FILENAME], where TYPE is "html", "text", "xml", or "console" and FILENAME is the filename (with
-        optional path) of the output report filename. If the report filename is
-        omitted, the default filename is used for the specified report type
-        ("CodeNarcReport.html" for "html" and "CodeNarcXmlReport.xml" for "xml"). If no
-        report option is specified, default to a single "html" report with the
-        default filename.
+        REPORT-TYPE[:FILENAME], where REPORT-TYPE is "html", "text", "xml", "json", "console"
+        or "consoleJson", and FILENAME is the filename (with  optional path) of the
+        output report filename. If the report filename is omitted, the default filename
+        is used for the specified report type ("CodeNarcReport.html" for "html",
+        "CodeNarcXmlReport.xml" for "xml", "CodeNarcJsonReport.json" for "json"). If no
+        report option is specified, default to a single "html" report with the default filename.
     -help
         Display the command-line help. If present, this must be the only command-line parameter.
   Example command-line invocations: