+++ src/main/groovy/org/codenarc/report/ReportWriterFactory.groovy
@@ -25,12 +25,13 @@ import org.codenarc.util.PropertyUtil
  * classpath) that implements the <code>org.codenarc.report.ReportWriter</code> interface.
  *
  * @author Chris Mair
+ * @author Nicolas Vuillamy
  */
 class ReportWriterFactory {
 
     ReportWriter getReportWriter(String type) {
         assert type
-        switch(type) {
+        switch (type) {
             case 'html': return new HtmlReportWriter()
             case 'sortable': return new SortableHtmlReportWriter()
             case 'xml': return new XmlReportWriter()
@@ -39,6 +40,8 @@ class ReportWriterFactory {
             case 'ide': def w = new IdeTextReportWriter(); w.writeToStandardOut = true; return w
             case 'inlineXml' : return new InlineXmlReportWriter()
             case 'baseline': return new BaselineXmlReportWriter()
+            case 'json': return new JsonReportWriter()
+            case 'consoleJson': def w = new JsonReportWriter(); w.writeToStandardOut = true; return w
         }
 
         def reportClass = getClass().classLoader.loadClass(type)
@@ -52,4 +55,5 @@ class ReportWriterFactory {
         }
         reportWriter
     }
+
 }